+++ Samples/CS/NewFunctions/RibbonMarkup.h
@@ -1,0 +1,54 @@
+// *****************************************************************************
+// * This is an automatically generated header file for UI Element definition  *
+// * resource symbols and values. Please do not modify manually.               *
+// *****************************************************************************
+
+#pragma once
+
+#define cmdApplication 2 
+#define cmdQat 3 
+#define cmdQatCustomize 4 
+#define cmdQatCustomize_LabelTitle_RESID 60001
+#define cmdQatCustomize_TooltipTitle_RESID 60002
+#define cmdTab1 5 
+#define cmdTab1_LabelTitle_RESID 60003
+#define cmdGroup1_1 6 
+#define cmdGroup1_1_LabelTitle_RESID 60004
+#define cmdDropDownColor 7 
+#define cmdDropDownColor_LabelTitle_RESID 60005
+#define cmdDropDownColor_TooltipTitle_RESID 60006
+#define cmdDropDownColor_TooltipDescription_RESID 60007
+#define cmdDropDownColor_SmallImages_RESID 60008
+#define cmdDropDownColor_LargeImages_RESID 60009
+#define cmdGroup1_2 8 
+#define cmdGroup1_2_LabelTitle_RESID 60010
+#define cmdFontPicker 9 
+#define cmdFontPicker_LabelTitle_RESID 60011
+#define cmdFontPicker_TooltipTitle_RESID 60012
+#define cmdFontPicker_TooltipDescription_RESID 60013
+#define cmdRecentItems 10 
+#define cmdRecentItems_LabelTitle_RESID 60014
+#define cmdRecentItems_LabelDescription_RESID 60015
+#define cmdRecentItems_TooltipTitle_RESID 60016
+#define cmdRecentItems_TooltipDescription_RESID 60017
+#define cmdButtonOpen 11 
+#define cmdButtonOpen_LabelTitle_RESID 60018
+#define cmdButtonOpen_TooltipTitle_RESID 60019
+#define cmdButtonOpen_TooltipDescription_RESID 60020
+#define cmdButtonOpen_SmallImages_RESID 60021
+#define cmdButtonOpen_LargeImages_RESID 60022
+#define cmdGroup1_3 12 
+#define cmdGroup1_3_LabelTitle_RESID 60023
+#define cmdInRibbon 13 
+#define cmdInRibbon_LabelTitle_RESID 60024
+#define cmdInRibbon_TooltipTitle_RESID 60025
+#define cmdInRibbon_TooltipDescription_RESID 60026
+#define cmdButtonHelp 14 
+#define cmdButtonHelp_LabelTitle_RESID 60027
+#define cmdButtonHelp_TooltipTitle_RESID 60028
+#define cmdButtonHelp_TooltipDescription_RESID 60029
+#define cmdGroup1_4 15 
+#define cmdGroup1_4_LabelTitle_RESID 60030
+#define cmdGroup1_4_LabelDescription_RESID 60031
+#define cmdComboBox 16 
+#define cmdComboBox_LabelTitle_RESID 60032